memperbarui kode

+++ database/migrations/2026_09_14_075016_create_orders_details_table.php
@@ -10,12 +10,22 @@ return new class extends Migration
     {
         Schema::create('order_details', function (Blueprint $table) {
             $table->id();
-            $table->foreignId('order_id')->constrained('orders')->onDelete('cascade');
-            $table->foreignId('service_id')->constrained('services')->onDelete('cascade');
-            $table->decimal('qty', 8, 2); 
+    
+            $table->foreignId('order_id')
+                  ->constrained('orders')
+                  ->onDelete('cascade');
+    
+            $table->foreignId('service_id')
+                  ->constrained('services')
+                  ->onDelete('cascade');
+    
+            $table->integer('qty');
             $table->decimal('subtotal', 12, 2);
             $table->timestamps();
         });
     }
+    
+    
 
 };
+
